Added require threads for windows
git-svn-id: https://svn.fhem.de/fhem/trunk@2655 2b470e98-0d58-463d-a4d8-8e2adae1ed80

+++ fhem/FHEM/Blocking.pm
@@ -93,7 +93,9 @@ sub
 BlockingKill($)
 {
   my $pid = shift;
-  Log 1, "Terminated $pid" if($pid && kill(9, $pid));
+  if($^O !~ m/Win/) {
+    Log 1, "Terminated $pid" if($pid && kill(9, $pid));
+  }
 }
 
 sub
@@ -103,6 +105,7 @@ BlockingExit($)
 
   if($^O =~ m/Win/) {
     close($client) if($client);
+    eval "require threads;";
     threads->exit();
 
   } else {
